requirements.txt
flask>=3.0
anthropic>=0.40
pypdf2>=3.0
openpyxl>=3.1
python-pptx>=0.6
python-docx>=1.0
Pillow>=10.0
